hybris: Use 64-bit usage variables where necessary
Some devices use private usage flags which are in the high bits
and got lost in the conversion. Currently only the low 32-bit usage
flags are transfered between processes, transfering the whole 64-bits,
would require changing the android_wlegl wayland protocol.

+++ hybris/egl/platforms/fbdev/fbdev_window.h
@@ -31,7 +31,7 @@ protected:
     FbDevNativeWindowBuffer(unsigned int width,
                             unsigned int height,
                             unsigned int format,
-                            unsigned int usage) ;
+                            uint64_t usage) ;
    virtual ~FbDevNativeWindowBuffer() ;
 
 protected:
