+++ hybris/platforms/common/server_wlegl.cpp
@@ -143,7 +143,7 @@ server_wlegl_get_server_buffer_handle(wl_client *client, wl_resource *res, uint3
 	// that we don't need to analyze the usage flags.
 	if (format == 0) format = HAL_PIXEL_FORMAT_RGBA_8888;
 
-	int r = hybris_gralloc_allocate(width, height, format, usage, &_handle, (uint32_t*)&_stride);
+	int r = hybris_gralloc_allocate(width, height, format, (uint32_t)usage, &_handle, (uint32_t*)&_stride);
 	if (r) {
 		HYBRIS_ERROR_LOG(SERVER_WLEGL, "failed to allocate buffer\n");
 		wl_resource_destroy(resource);
