+++ hybris/platforms/common/server_wlegl_buffer.cpp
@@ -60,7 +60,7 @@ server_wlegl_buffer_create(wl_client *client,
 			   int32_t height,
 			   int32_t stride,
 			   int32_t format,
-			   int32_t usage,
+			   int64_t usage,
 			   buffer_handle_t handle,
 			   server_wlegl *wlegl)
 {
@@ -90,7 +90,7 @@ server_wlegl_buffer_create_server(wl_client *client,
 			   int32_t height,
 			   int32_t stride,
 			   int32_t format,
-			   int32_t usage,
+			   int64_t usage,
 			   buffer_handle_t handle,
 			   server_wlegl *wlegl)
 {
